Update fr.yaml
Strictly speaking regions at level 4 are not "states".

States could be for level 3 which designates the contiguous territorial areas : one for the Metropole, one for each oversea department or collectivity. However France is not federal, it legally has a single "state" (including the very specific status of New Caledonia where lot of powers are delegated, including a plain government acting in most sectors except defense and security, nationality, and high juridictions, so New Caledonia is not lkegally a state and has no separate constitution, it is regulated by the French Constitution itself with a special reserve allowing it to become independant after several referunda)

Levels 6 and 7 are very clearly distinct in France and purely hierarchical. They absolutely require a distinction, they are not equivalent: 6 for departments (collectivity definition), 7 for their "arrondissements" (not the same as municipal arrondissements in Paris, Lyon and Marseille Only). There's an exception with level 5 for the department of Rhone (not the collectivity, but the area managed at prefectoral level containing the Department of Rhone, and the Metropole de Lyon, both at level 6).

So level 6 could be managed like "state district", while level 7 behaves more like "county".

Note finally that level 9 is not just for city districts, it is used also for communes that are are not plain municipalities (i.e. still considered communes with distinct territories and some distinctive budgets, and a local "delegated" mayor, but working inside the same merged municipal council (for other grouped communes, this is in fact just like in municipal arrondissements of Paris, Lyon, and Marseille, except that these 3 cities have an elected local council below the municipal level, and that the 16 arrondissements of Marseille are grouped by two into "sectors" for electing their mayor, while each municipal arrondissements still has its own council): there are several such statuses for non-autonomous communes (which were plain municipalities in the past and can decide tyo return to a full municipality: they jsut don't have an autonomous fiscality and budget and cooperate in the same coucil, that suballocates to each member commune a small dedicated budget for their limited local missions).
Only communes that are fully merged will no longer have a level 9 subdivision (but they are now a minority, full merging is no longer necessary with the new status of "commune nouvelle" in France).

So this level 9 in France is exactly similar to the case of "municipal sections" in Belgium (most communes in Belgium have grouped in such way, in France there are lot of undivided communes but the trend is growing with "communes nouvelles" because of state incentive).